Check in Cargo.lock
This is the Cargo.lock file created by running cargo update in
the working tree of the baseline commit in the master branch.
The GitHub workflow defined in update-deps.yml keeps this commit
up to date with the dependencies checked against the head of the master
branch; whenever the Cargo.lock contents change, this commit is rebased.